+++ spec/image_size_spec.rb
@@ -196,6 +196,9 @@ describe ImageSize do
       context 'fetching from webserver' do
         let(:file_url){ @server.base_url + path }
 
+        before{ ImageSize.chunk_size = 64 }
+        after{ ImageSize.chunk_size = nil }
+
         context 'supporting range' do
           context 'without redirects' do
             it 'gets format and dimensions' do
@@ -222,6 +225,17 @@ describe ImageSize do
                   ImageSize.url("#{file_url}?redirect=6")
                 end
               end.to raise_error(/Too many redirects/)
+            end
+          end
+
+          context 'with unknown file size' do
+            let(:attributes){ super().merge(byte_size: file_size.zero? ? 0 : nil) }
+
+            it 'gets format and dimensions' do
+              image_size = retry_on Timeout::Error do
+                ImageSize.url("#{file_url}?unknown_file_size")
+              end
+              expect(image_size).to have_attributes(attributes)
             end
           end
         end